Added TestStartup (TestStartup/Test_Startup.xml, etc.) Entity.
Fixed a problem with Viewport implementation - size and position stored were absolute when it makes more sense for them to be relative.

Entity and Camera are now garbage collected.

Fixed some pretty serious threading related bugs in ResourceManager.

Incomplete implementation of player-ID synchronisation.

Fixed a dopey bug with both CScriptString and ScriptVector (not returning a valid type ID because I didn't read the AngelScript manual and was attempting to use functionality that didn't exist to get said ID).

Struggled with a conflict between abstraction and simplicity, and the meaning of the latter, with regard to Network interface. Somewhat apathetic, since all but the highest level networking issues are not particularly interesting to me.